+++ src/main/java/com/kharis/todo/repository/TodoSpringDataJpaRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.kharis.todo.model.Todo;
 
-public interface TodoSpringDataJpaRepo extends JpaRepository<Todo, Integer> {
+public interface TodoSpringDataJpaRepo extends JpaRepository<Todo, Long> {
 
 }
